+++ src/main/java/com/codenear/butterfly/consent/infrastructure/ConsentRepository.java
@@ -1,5 +1,6 @@
-package com.codenear.butterfly.consent.domain;
+package com.codenear.butterfly.consent.infrastructure;
 
+import com.codenear.butterfly.consent.domain.Consent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
